third_party/blink/web_tests/animations/responsive/interpolation/font-size-adjust-responsive-expected.txt

CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
CONSOLE WARNING: Invalid keyframe value for property fontSizeAdjust: ic-height 1.0
This is a testharness.js-based test.
[FAIL] font-size-adjust interpolation responds to changes in inherited value 60
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 61
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 62
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 63
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 64
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 65
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 66
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 67
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 68
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 69
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 70
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 71
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 132
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 133
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 134
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 135
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 136
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 137
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 138
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 139
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 140
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 141
  assert_true: CSS.supports font-size-adjust ic-height 2.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 142
  assert_true: CSS.supports font-size-adjust ic-height 0.0 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 143
  assert_true: CSS.supports font-size-adjust ic-height 0.5 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 264
  assert_true: CSS.supports font-size-adjust ic-height 0.625 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 265
  assert_true: CSS.supports font-size-adjust ic-height 0.875 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 266
  assert_true: CSS.supports font-size-adjust ic-height 1.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 267
  assert_true: CSS.supports font-size-adjust ic-height 1.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 268
  assert_true: CSS.supports font-size-adjust ic-height 0.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 269
  assert_true: CSS.supports font-size-adjust ic-height 0.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 270
  assert_true: CSS.supports font-size-adjust ic-height 1.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 271
  assert_true: CSS.supports font-size-adjust ic-height 1.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 272
  assert_true: CSS.supports font-size-adjust ic-height 0.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 273
  assert_true: CSS.supports font-size-adjust ic-height 0.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 274
  assert_true: CSS.supports font-size-adjust ic-height 0.625 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 275
  assert_true: CSS.supports font-size-adjust ic-height 0.875 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 276
  assert_true: CSS.supports font-size-adjust ic-height 0.875 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 277
  assert_true: CSS.supports font-size-adjust ic-height 0.625 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 278
  assert_true: CSS.supports font-size-adjust ic-height 1.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 279
  assert_true: CSS.supports font-size-adjust ic-height 1.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 280
  assert_true: CSS.supports font-size-adjust ic-height 0.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 281
  assert_true: CSS.supports font-size-adjust ic-height 0.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 282
  assert_true: CSS.supports font-size-adjust ic-height 1.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 283
  assert_true: CSS.supports font-size-adjust ic-height 1.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 284
  assert_true: CSS.supports font-size-adjust ic-height 0.75 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 285
  assert_true: CSS.supports font-size-adjust ic-height 0.25 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 286
  assert_true: CSS.supports font-size-adjust ic-height 0.875 expected true got false
[FAIL] font-size-adjust interpolation responds to changes in inherited value 287
  assert_true: CSS.supports font-size-adjust ic-height 0.625 expected true got false
Harness: the test ran to completion.
